Move keydown('n') decorator
